crisp_threat_intel/requirements/development.txt
-r base.txt

# Development dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
factory-boy>=3.2.0
ipython>=8.0.0
django-debug-toolbar>=3.2.0